+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["setuptools>=61.0"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "fipy"
+name = "fippy"
 version = "2023.0.0"
 dependencies = [
     "enlighten",
