+++ src/fippy/backend/categorical/__init__.py
