+++ src/fippy/backend/categorical/categorical_estimator.py
