+++ src/fippy/backend/cnf/transforms.py
