+++ src/fippy/backend/gaussian/gaussian_estimator.py
