+++ src/fippy/backend/goodness_of_fit.py
