+++ src/fippy/examples/chains.py
