+++ src/fippy/examples/ii_paper.py
