+++ src/fippy/explanation/decomposition.py
