+++ src/fippy/plots/_barplot.py
