+++ src/fippy/plots/_barplot_multiple.py
