+++ src/fippy/plots/_snsstyle.py
