+++ src/fippy/samplers/gaussian.py
