+++ src/fippy/samplers/mdn.py
