+++ src/fippy/samplers/sequential.py
